Added payment method constants for giropay and eps
While I was at it, I figured it couldn't hurt to alphabetically sort the payment method constants. Furthermore, I added the `giftcard` method, as it was missing.`

I updated the README to include the appropriate text for the new payment methods, but as the links to the respective payment methods are not present yet, I could not add links yet.