Refactor to export a better type definition file.
I've made one breaking change: I've replaced the addVersionString method for a versionStrings option.

I think we had an addVersionString method because the PHP API has it, but the PHP API also has a setApiKey method. If we are breaking consistency with the PHP API anyway, I prefer putting the version strings in the options. It makes for a cleaner interface, and version strings probably don't change frequently to justify having a method.