Now also exporting a named createMollieClient function.
If this library is used CJS require-style instead of ES import-style, one would otherwise have to type '.default' to access the function. It was previously possible to export that function directly instead of encapsulating it in an object, but that is no longer the case as we are now also exporting enums and things.