Remove husky dependency (which hasn't been active for a while).
I personally don't find Husky useful in the way we were using it before. We have Travis CI in place to ensure any failing tests don't go unnoticed, even if the developer somehow forgets to run the tests themselves. The downside of having Husky trigger Jest, is that Jest tests the working tree ‒ not the index. If you try to commit some but not all changes (in other words, you don't do "git add ."), Husky will complain if the working tree does not pass the tests; even if the index (the thing you are trying to commit) does.

If you feel we should reintroduce Husky, please reach out to me.