Remove endpoint aliases.
These endpoint aliases were deprecated in https://github.com/mollie/mollie-api-node/pull/234. The first version to include this deprecation is 3.6.0, released on 10 February 2022.

In the era of TypeScript, I feel the confusion caused by having two (or sometimes three) separate functions in the type definition outweighs the backwards compatibility. Consumers of an API expect every part of it to be significant.